[chromium] Bypass the shadow texture copy for accelerated canvas when running the compositor in single threaded mode. https://bugs.webkit.org/show_bug.cgi?id=75146
Source/WebCore:

The texture copy fails on Windows as glCopyTexImage2D() doesn't
support BGRA source textures.
This change also modifies Canvas2DLayerChromium::updateCompositorResources
to call glCopyTexSubImage2D() instead of glCopyTexImage2D() so that
the copy can work with texture allocated via the glTexStorage2D
extension.

Reviewed by James Robinson.

Tests: Canvas2DLayerChromiumTest.cpp

* platform/graphics/chromium/Canvas2DLayerChromium.cpp:
(WebCore::Canvas2DLayerChromium::Canvas2DLayerChromium):
(WebCore::Canvas2DLayerChromium::~Canvas2DLayerChromium):
(WebCore::Canvas2DLayerChromium::paintContentsIfDirty):
(WebCore::Canvas2DLayerChromium::setTextureManager):
(WebCore::Canvas2DLayerChromium::updateCompositorResources):
(WebCore::Canvas2DLayerChromium::pushPropertiesTo):
(WebCore::Canvas2DLayerChromium::unreserveContentsTexture):
(WebCore::Canvas2DLayerChromium::cleanupResources):
* platform/graphics/chromium/Canvas2DLayerChromium.h:

Source/WebKit/chromium:

Modified Canvas2DLayerChromium unit test to test canvas with
both threaded and non-threaded compositing.

Reviewed by James Robinson.

* tests/CCSchedulerTestCommon.h:
(WebKitTests::FakeCCThread::threadID):
* tests/Canvas2DLayerChromiumTest.cpp:
(WebCore::Canvas2DLayerChromiumTest::setTextureManager):
(WebCore::Canvas2DLayerChromiumTest::fullLifecycleTest):
(WebCore::TEST_F):


Canonical link: https://commits.webkit.org/92331@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@104044 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/chromium/Canvas2DLayerChromium.cpp
@@ -58,13 +58,16 @@ Canvas2DLayerChromium::Canvas2DLayerChromium(GraphicsContext3D* context, const I
     , m_size(size)
     , m_backTextureId(0)
     , m_fbo(0)
+    , m_useDoubleBuffering(CCProxy::hasImplThread())
 {
-    GLC(m_context, m_fbo = m_context->createFramebuffer());
+    if (m_useDoubleBuffering)
+        GLC(m_context, m_fbo = m_context->createFramebuffer());
 }
 
 Canvas2DLayerChromium::~Canvas2DLayerChromium()
 {
-    GLC(m_context, m_context->deleteFramebuffer(m_fbo));
+    if (m_useDoubleBuffering && m_fbo)
+       GLC(m_context, m_context->deleteFramebuffer(m_fbo));
 }
 
 void Canvas2DLayerChromium::setTextureId(unsigned textureId)
@@ -92,7 +95,8 @@ void Canvas2DLayerChromium::paintContentsIfDirty()
     if (!drawsContent())
         return;
 
-    m_frontTexture->reserve(m_size, GraphicsContext3D::RGBA);
+    if (m_useDoubleBuffering)
+        m_frontTexture->reserve(m_size, GraphicsContext3D::RGBA);
 
     if (!needsDisplay())
         return;
@@ -121,7 +125,7 @@ void Canvas2DLayerChromium::setLayerTreeHost(CCLayerTreeHost* host)
 
 void Canvas2DLayerChromium::setTextureManager(TextureManager* textureManager)
 {
-    if (textureManager)
+    if (textureManager && m_useDoubleBuffering)
         m_frontTexture = ManagedTexture::create(textureManager);
     else
         m_frontTexture.clear();
@@ -129,14 +133,16 @@ void Canvas2DLayerChromium::setTextureManager(TextureManager* textureManager)
 
 void Canvas2DLayerChromium::updateCompositorResources(GraphicsContext3D* context, CCTextureUpdater& updater)
 {
-    if (!m_backTextureId || !m_frontTexture->isValid(m_size, GraphicsContext3D::RGBA))
+    if (!m_backTextureId || !m_frontTexture || !m_frontTexture->isValid(m_size, GraphicsContext3D::RGBA))
         return;
 
     m_frontTexture->bindTexture(context, updater.allocator());
 
     GLC(context, context->bindFramebuffer(GraphicsContext3D::FRAMEBUFFER, m_fbo));
     GLC(context, context->framebufferTexture2D(GraphicsContext3D::FRAMEBUFFER, GraphicsContext3D::COLOR_ATTACHMENT0, GraphicsContext3D::TEXTURE_2D, m_backTextureId, 0));
-    GLC(context, context->copyTexImage2D(GraphicsContext3D::TEXTURE_2D, 0, GraphicsContext3D::RGBA, 0, 0, m_size.width(), m_size.height(), 0));
+    // FIXME: The copy operation will fail if the m_backTexture is allocated as BGRA since glCopyTex(Sub)Image2D doesn't
+    //        support the BGRA format. See bug https://bugs.webkit.org/show_bug.cgi?id=75142
+    GLC(context, context->copyTexSubImage2D(GraphicsContext3D::TEXTURE_2D, 0, 0, 0, 0, 0, m_size.width(), m_size.height()));
     GLC(context, context->bindFramebuffer(GraphicsContext3D::FRAMEBUFFER, 0));
     GLC(context, context->flush());
 }
@@ -146,17 +152,22 @@ void Canvas2DLayerChromium::pushPropertiesTo(CCLayerImpl* layer)
     CanvasLayerChromium::pushPropertiesTo(layer);
 
     CCCanvasLayerImpl* canvasLayer = static_cast<CCCanvasLayerImpl*>(layer);
-    canvasLayer->setTextureId(m_frontTexture->textureId());
+    if (m_useDoubleBuffering)
+        canvasLayer->setTextureId(m_frontTexture->textureId());
+    else
+        canvasLayer->setTextureId(m_backTextureId);
 }
 
 void Canvas2DLayerChromium::unreserveContentsTexture()
 {
-    m_frontTexture->unreserve();
+    if (m_useDoubleBuffering)
+        m_frontTexture->unreserve();
 }
 
 void Canvas2DLayerChromium::cleanupResources()
 {
-    m_frontTexture.clear();
+    if (m_useDoubleBuffering)
+        m_frontTexture.clear();
 }
 
 }
